Enhance ECS API with automatic dependency injection and serialisation (only NBT right now)
This includes separating and exposing the entire serialisation framework previously used exclusively for the networking implementation. The Serialisation API will therefore require a cleanup eventually.
The dependency injection is currently limited to the component holder (i.e. an instance of Allomancer) and the corresponding entity.
A general purpose implementation of Allomancer for entities was introduced, easily reusable by means of Forge capabilities.